[#2] added vertex ai implementation
