--- vertexai/pkg.go
@@ -1,9 +1,0 @@
-package vertexai
-
-import "oss.nandlabs.io/golly/l3"
-
-var logger = l3.Get()
-
-func init() {
-	logger.Info("vertex init")
-}
